reduced components
