--- igenio/src/components/GuestMain.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react';
-import Header from './components/partials/Header';
-
-
-class Home extends Component {
-  Render() {
-    Return (
-      <h1>INGENIO</h1>
-      <RandomIdea />
-
-      )
-  }
-}
-
-export default Home;
