--- igenio/src/components/RandomIdea.js
@@ -1,1 +1,0 @@
-//Class that renders random tweet from Twitter search api
