+++ igenio/src/components/RegisterForm.jsx
@@ -1,40 +1,74 @@
 import React, { Component } from 'react';
+import axios from 'axios';
 
 class RegisterForm extends Component{
+  constructor() {
+    super();
+    this.state = {
+      apiData: [],
+      isLoggedIn: false,
+      registerEmail: '',
+      registerUserName: '',
+      registerPassword: '',
+    };
+  }
 
+  componentDidMount() {
+    console.log('compontent mounted');
+  }
 
-  render(){
+  onSubmitRegister(e) {
+    e.preventDefault();
+    console.log(this.state);
+    axios.post('/api/AEMIRO', {
+      registerEmail: this.state.registerEmail,
+      registerUserName: this.state.registerUserName,
+      registerPassword: this.state.registerPassword,
+    });
+  }
+
+  handleInput(e) {
+    const name = e.target.name;
+    const value = e.target.value;
+    this.setState({
+      [name]: value,
+    });
+  }
+
+  render() {
     return (
       <div>
-      <a href="/auth/twitter"><button>Login With Twitter</button></a>
-      <form onSubmit={this.onSubmitRegister} className="register-form">
-        <input
-          name="registerEmail"
-          type="email"
-          placeholder="Your Email"
-          value={this.props.registerEmail}
-          onChange={this.props.handleInput}
+        <a href="/auth/twitter"><button>Login With Twitter</button></a>
+        <form onSubmit={this.onSubmitRegister} className="register-form">
+          <input
+            name="registerEmail"
+            type="email"
+            placeholder="Your Email"
+            value={this.state.registerEmail}
+            onChange={this.handleInput}
           />
-      <br />
-           <input
-           name="registerUserName"
-          type="text"
-          placeholder="Your Username"
-          value={this.props.registerUserName}
-          onChange={this.props.handleInput}
+          <br />
+          <input
+            name="registerUserName"
+            type="text"
+            placeholder="Your Username"
+            value={this.state.registerUserName}
+            onChange={this.handleInput}
           />
-      <br />
-           <input
-           name="registerPassword"
-          type="password"
-          placeholder="Your Password"
-          value={this.props.registerpassword}
-          onChange={this.props.handleInput}
+          <br />
+          <input
+            name="registerPassword"
+            type="password"
+            placeholder="Your Password"
+            value={this.state.registerpassword}
+            onChange={this.handleInput}
           />
-      <br />
-         <button type="submit">Submit</button>
-      </form>
+          <br />
+          <button type="submit">Submit</button>
+        </form>
       </div>
-      )
+    );
   }
 }
+
+export default RegisterForm;
