--- igenio/src/components/UserProfile.js
@@ -1,2 +1,0 @@
-//Class that renders User data from DB and a settings button that allows
-//them to edit and update their user information.
