+++ igenio/src/components/home_2.jsx
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+
+class Home_2 extends Component {
+
+
+  Render() {
+    Return (
+      <div>
+        <IdeasTab />
+        <CategoryButtons />
+      </div>,
+
+    );
+  }
+}
+
+export default Home_2;
